+++ client/view/js/index.js
@@ -618,7 +618,7 @@ let NormalizedEvent=class NormalizedEvent {
         return this.event.timeStamp;
     }
     get pointerType() {
-        if (this._event instanceof TouchEvent)
+        if ('TouchEvent' in window && this._event instanceof TouchEvent)
             return "touch";
         return this.getProp("pointerType");
     }
@@ -626,7 +626,7 @@ let NormalizedEvent=class NormalizedEvent {
         return this.getProp("button");
     }
     get isTouch() {
-        if (this._event instanceof TouchEvent)
+        if ('TouchEvent' in window && this._event instanceof TouchEvent)
             return true;
         return this._event.pointerType == "touch";
     }
@@ -1970,7 +1970,7 @@ let PressManager=class PressManager {
         return touch.pointerId;
     }
     registerEvent(ev) {
-        if (ev instanceof TouchEvent) {
+        if ('TouchEvent' in window && ev instanceof TouchEvent) {
             for (let touch of ev.targetTouches) {
                 const id = this.identifyEvent(touch);
                 if (this.pointersRecord[id]) {
@@ -1991,7 +1991,7 @@ let PressManager=class PressManager {
     }
     unregisterEvent(ev) {
         let result = true;
-        if (ev instanceof TouchEvent) {
+        if ('TouchEvent' in window && ev instanceof TouchEvent) {
             for (let touch of ev.changedTouches) {
                 const id = this.identifyEvent(touch);
                 if (!this.pointersRecord[id]) {